+++ custom_addons/odoo_ai_ops/models/__init__.py
@@ -1,2 +1,9 @@
 # -*- coding: utf-8 -*-
-from . import ai_ops_inventory, ai_ops_order_risk, ai_ops_settings, ai_ops_task
+from . import (
+    ai_ops_inventory,
+    ai_ops_order_intake,
+    ai_ops_order_risk,
+    ai_ops_settings,
+    ai_ops_task,
+    sale_order,
+)